// backend_vibeazy/src/interfaces/flutterwave-interface.ts
export interface WebhookDataProd {
  event: 'charge.completed',
  data: {
    id: number,
    tx_ref: string,
    reference: string
    flw_ref: string,
    device_fingerprint: string,
    amount: 30,
    currency: 'NGN',
    charged_amount: number,
    app_fee: number,
    merchant_fee: number,
    processor_response: string,
    auth_model:string,
    ip: string,
    narration:string,
    status: string,
    payment_type: string,
    created_at: string,
    account_id: 683269,
    customer: {
      id: 994192822,
      name: string,
      phone_number: string,
      email: string,
      created_at: string
    }
  },
  'event.type': 'BANK_TRANSFER_TRANSACTION' | "Transfer"
}


export interface ICardDetails {
    first_6digits: string;
    last_4digits: string;
    issuer: string;
    country: string;
    type: string;
    token: string;
    expiry: string;
  }

  
export interface IFlwTransactionVerificationData {
    id: number;
    tx_ref: string;
    flw_ref: string;
    device_fingerprint: string;
    amount: number;
    currency: string;
    charged_amount: number;
    app_fee: number;
    merchant_fee: number;
    processor_response: string;
    auth_model: string;
    ip: string;
    narration: string;
    status: string;
    payment_type: string;
    created_at: string; // DateTime in ISO format
    account_id: number;
    card?: ICardDetails;
    plan: number;
    amount_settled: number;
    customer: CustomerDetails;
  }

interface CustomerDetails {
    id: number;
    name: string;
    phone_number: string;
    email: string;
    created_at: string; // DateTime in ISO format
}